% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixGenerics-package.R
\name{internal-helpers}
\alias{internal-helpers}
\alias{normarg_center}
\title{Internal helpers}
\usage{
normarg_center(center, n, what)
}
\description{
Not for end users
}
\keyword{internal}
\keyword{utilities}
