% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MatrixGenerics-package.R
\docType{class}
\name{MatrixGenerics-package}
\alias{MatrixGenerics-package}
\alias{class:matrix_OR_array_OR_table_OR_numeric}
\alias{matrix_OR_array_OR_table_OR_numeric-class}
\alias{matrix_OR_array_OR_table_OR_numeric}
\title{The MatrixGenerics package}
\description{
The \pkg{MatrixGenerics} package defines S4 generic summary statistic
functions that operate on matrix-Like objects.
}
