% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowAlls.R
\name{rowAlls}
\alias{rowAlls}
\alias{rowAlls,matrix-method}
\alias{rowAlls,numeric-method}
\alias{rowAlls,array-method}
\alias{colAlls}
\alias{colAlls,matrix-method}
\alias{colAlls,numeric-method}
\alias{colAlls,array-method}
\title{Check if all elements in a row (column) of a matrix-like object are equal to a value}
\usage{
rowAlls(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE, ...)

\S4method{rowAlls}{matrix}(
  x,
  rows = NULL,
  cols = NULL,
  value = TRUE,
  na.rm = FALSE,
  dim. = dim(x)
)

\S4method{rowAlls}{numeric}(
  x,
  rows = NULL,
  cols = NULL,
  value = TRUE,
  na.rm = FALSE,
  dim. = dim(x)
)

\S4method{rowAlls}{array}(
  x,
  rows = NULL,
  cols = NULL,
  value = TRUE,
  na.rm = FALSE,
  dim. = dim(x)
)

colAlls(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE, ...)

\S4method{colAlls}{matrix}(
  x,
  rows = NULL,
  cols = NULL,
  value = TRUE,
  na.rm = FALSE,
  dim. = dim(x)
)

\S4method{colAlls}{numeric}(
  x,
  rows = NULL,
  cols = NULL,
  value = TRUE,
  na.rm = FALSE,
  dim. = dim(x)
)

\S4method{colAlls}{array}(
  x,
  rows = NULL,
  cols = NULL,
  value = TRUE,
  na.rm = FALSE,
  dim. = dim(x)
)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows, cols}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{value}{the value to search for.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, essential when x is a
\code{\link[base]{numeric}} vector.}
}
\value{
Returns a \code{\link[base]{logical}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Check if all elements in a row (column) of a matrix-like object are equal to a value
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or
\code{\link[base]{numeric}} call \code{matrixStats::\link[matrixStats]{rowAlls}}
/ \code{matrixStats::\link[matrixStats]{colAlls}}.
}
\examples{
  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowAlls(mat)
  colAlls(mat)
  
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowAlls}()} and
  \code{matrixStats::\link[matrixStats]{colAlls}()} which are used when
  the input is a \code{matrix} or \code{numeric} vector.
\item For checks if any element is equal to a value, see \code{\link{rowAnys}()}.
}
}
