% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowAnys.R
\name{rowAnys}
\alias{rowAnys}
\alias{rowAnys,matrix-method}
\alias{rowAnys,numeric-method}
\alias{rowAnys,array-method}
\alias{colAnys}
\alias{colAnys,matrix-method}
\alias{colAnys,numeric-method}
\alias{colAnys,array-method}
\title{Check if any elements in a row (column) of a matrix-like object is equal to
a value}
\usage{
rowAnys(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE, ...)

\S4method{rowAnys}{matrix}(x, rows = NULL, cols = NULL, value = TRUE,
  na.rm = FALSE, dim. = dim(x), ...)

\S4method{rowAnys}{numeric}(x, rows = NULL, cols = NULL, value = TRUE,
  na.rm = FALSE, dim. = dim(x), ...)

\S4method{rowAnys}{array}(x, rows = NULL, cols = NULL, value = TRUE,
  na.rm = FALSE, dim. = dim(x), ...)

colAnys(x, rows = NULL, cols = NULL, value = TRUE, na.rm = FALSE, ...)

\S4method{colAnys}{matrix}(x, rows = NULL, cols = NULL, value = TRUE,
  na.rm = FALSE, dim. = dim(x), ...)

\S4method{colAnys}{numeric}(x, rows = NULL, cols = NULL, value = TRUE,
  na.rm = FALSE, dim. = dim(x), ...)

\S4method{colAnys}{array}(x, rows = NULL, cols = NULL, value = TRUE,
  na.rm = FALSE, dim. = dim(x), ...)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows, cols}{A \code{\link{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link{NULL}}, no subsetting is
done.}

\item{value}{The value to search for.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link{NA}}s
are excluded first, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}

\item{dim.}{An \code{\link{integer}} \code{\link{vector}} of length two
specifying the dimension of \code{x}, essential when x is a
\code{\link{numeric}} vector. Note, that this is not a generic
argument and not all methods need provide it.}
}
\value{
Returns a \code{\link{logical}} \code{\link{vector}} of length N (K).
}
\description{
Check if any elements in a row (column) of a matrix-like object is equal to
a value.
}
\details{
The S4 methods for \code{x} of type \code{\link{matrix}},
\code{\link{array}}, or \code{\link{numeric}} call
\code{matrixStats::rowAnys}
/ \code{matrixStats::colAnys}.
}
\examples{
  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowAnys(mat)
  colAnys(mat)
  
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats:rowAlls]{rowAnys}()} and
\code{matrixStats::\link[matrixStats:rowAlls]{colAnys}()} which are used
when the input is a \code{matrix} or \code{numeric} vector.
\item \code{base::\link{any}()}.
}
}
