% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowCollapse.R
\name{rowCollapse}
\alias{rowCollapse}
\alias{rowCollapse,matrix-method}
\alias{rowCollapse,numeric-method}
\alias{rowCollapse,array-method}
\alias{colCollapse}
\alias{colCollapse,matrix-method}
\alias{colCollapse,numeric-method}
\alias{colCollapse,array-method}
\title{Extract one cell from each row (column) of a matrix-like object}
\usage{
rowCollapse(x, idxs, rows = NULL, ...)

\S4method{rowCollapse}{matrix}(x, idxs, rows = NULL, dim. = dim(x))

\S4method{rowCollapse}{numeric}(x, idxs, rows = NULL, dim. = dim(x))

\S4method{rowCollapse}{array}(x, idxs, rows = NULL, dim. = dim(x))

colCollapse(x, idxs = idxs, cols = NULL, ...)

\S4method{colCollapse}{matrix}(x, idxs, cols = NULL, dim. = dim(x))

\S4method{colCollapse}{numeric}(x, idxs, cols = NULL, dim. = dim(x))

\S4method{colCollapse}{array}(x, idxs, cols = NULL, dim. = dim(x))
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{idxs}{An index \code{\link[base]{vector}} with the position to extract.
It is recycled to match the number of rows (column)}

\item{rows, cols}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{...}{Additional arguments passed to specific methods.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, essential when x is a
\code{\link[base]{numeric}} vector.
Note, that this is not a generic argument and not all implementation of this function
have to provide it.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Extract one cell from each row (column) of a matrix-like object
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or
\code{\link[base]{numeric}} call \code{matrixStats::\link[matrixStats]{rowCollapse}}
/ \code{matrixStats::\link[matrixStats]{colCollapse}}.
}
\examples{
  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowCollapse(mat, idxs = 2)
  rowCollapse(mat, idxs = c(1,1,2,3,2))
  
  colCollapse (mat, idxs = 4)

}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowCollapse}()} and
\code{matrixStats::\link[matrixStats]{colCollapse}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
}
}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
