% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowCummins.R
\name{rowCummins}
\alias{rowCummins}
\alias{rowCummins,matrix-method}
\alias{rowCummins,numeric-method}
\alias{rowCummins,array-method}
\alias{colCummins}
\alias{colCummins,matrix-method}
\alias{colCummins,numeric-method}
\alias{colCummins,array-method}
\title{Calculates the cumulative minima for each row (column) of a matrix-like
object}
\usage{
rowCummins(x, rows = NULL, cols = NULL, ...)

\S4method{rowCummins}{matrix}(x, rows = NULL, cols = NULL, dim. = dim(x), ...)

\S4method{rowCummins}{numeric}(x, rows = NULL, cols = NULL, dim. = dim(x), ...)

\S4method{rowCummins}{array}(x, rows = NULL, cols = NULL, dim. = dim(x), ...)

colCummins(x, rows = NULL, cols = NULL, ...)

\S4method{colCummins}{matrix}(x, rows = NULL, cols = NULL, dim. = dim(x), ...)

\S4method{colCummins}{numeric}(x, rows = NULL, cols = NULL, dim. = dim(x), ...)

\S4method{colCummins}{array}(x, rows = NULL, cols = NULL, dim. = dim(x), ...)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows, cols}{A \code{\link{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link{NULL}}, no subsetting is
done.}

\item{...}{Additional arguments passed to specific methods.}

\item{dim.}{An \code{\link{integer}} \code{\link{vector}} of length two
specifying the dimension of \code{x}, essential when x is a
\code{\link{numeric}} vector. Note, that this is not a generic
argument and not all methods need provide it.}
}
\value{
Returns a \code{\link{numeric}} \code{\link{matrix}}with the same
dimensions as \code{x}.
}
\description{
Calculates the cumulative minima for each row (column) of a matrix-like
object.
}
\details{
The S4 methods for \code{x} of type \code{\link{matrix}},
\code{\link{array}}, or \code{\link{numeric}} call
\code{matrixStats::\link[matrixStats]{rowCummins}}
/ \code{matrixStats::\link[matrixStats:rowCummins]{colCummins}}.
}
\examples{
  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowCummins(mat)
  colCummins(mat)
  
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowCummins}()} and
\code{matrixStats::\link[matrixStats:rowCummins]{colCummins}()} which are
used when the input is a \code{matrix} or \code{numeric} vector.
\item For single minimum estimates, see \code{\link{rowMins}()}.
\item \code{base::\link{cummin}()}.
}
}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
