% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowCumprods.R
\name{rowCumprods}
\alias{rowCumprods}
\alias{rowCumprods,matrix-method}
\alias{rowCumprods,numeric-method}
\alias{rowCumprods,array-method}
\alias{colCumprods}
\alias{colCumprods,matrix-method}
\alias{colCumprods,numeric-method}
\alias{colCumprods,array-method}
\title{Calculates the cumulative maxima for each row (column) of a matrix-like object}
\usage{
rowCumprods(x, rows = NULL, cols = NULL, ...)

\S4method{rowCumprods}{matrix}(x, rows = NULL, cols = NULL, dim. = dim(x))

\S4method{rowCumprods}{numeric}(x, rows = NULL, cols = NULL, dim. = dim(x))

\S4method{rowCumprods}{array}(x, rows = NULL, cols = NULL, dim. = dim(x))

colCumprods(x, rows = NULL, cols = NULL, ...)

\S4method{colCumprods}{matrix}(x, rows = NULL, cols = NULL, dim. = dim(x))

\S4method{colCumprods}{numeric}(x, rows = NULL, cols = NULL, dim. = dim(x))

\S4method{colCumprods}{array}(x, rows = NULL, cols = NULL, dim. = dim(x))
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows, cols}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{...}{Additional arguments passed to specific methods.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, essential when x is a
\code{\link[base]{numeric}} vector.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{matrix}} with the
same dimensions as \code{x}.
}
\description{
Calculates the cumulative maxima for each row (column) of a matrix-like object
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or
\code{\link[base]{numeric}} call \code{matrixStats::\link[matrixStats]{rowCumprods}}
/ \code{matrixStats::\link[matrixStats]{colCumprods}}.
}
\examples{
  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowCumprods(mat)
  colCumprods(mat)
  
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowCumprods}()} and
  \code{matrixStats::\link[matrixStats]{colCumprods}()} which are used when
  the input is a \code{matrix} or \code{numeric} vector.
}
}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
