% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowCumsums.R
\name{rowCumsums}
\alias{rowCumsums}
\alias{rowCumsums,matrix-method}
\alias{rowCumsums,numeric-method}
\alias{rowCumsums,array-method}
\alias{colCumsums}
\alias{colCumsums,matrix-method}
\alias{colCumsums,numeric-method}
\alias{colCumsums,array-method}
\title{Calculates the cumulative sum for each row (column) of a matrix-like object}
\usage{
rowCumsums(x, rows = NULL, cols = NULL, ...)

\S4method{rowCumsums}{matrix}(x, rows = NULL, cols = NULL, dim. = dim(x))

\S4method{rowCumsums}{numeric}(x, rows = NULL, cols = NULL, dim. = dim(x))

\S4method{rowCumsums}{array}(x, rows = NULL, cols = NULL, dim. = dim(x))

colCumsums(x, rows = NULL, cols = NULL, ...)

\S4method{colCumsums}{matrix}(x, rows = NULL, cols = NULL, dim. = dim(x))

\S4method{colCumsums}{numeric}(x, rows = NULL, cols = NULL, dim. = dim(x))

\S4method{colCumsums}{array}(x, rows = NULL, cols = NULL, dim. = dim(x))
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows, cols}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{...}{Additional arguments passed to specific methods.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, essential when x is a
\code{\link[base]{numeric}} vector.
Note, that this is not a generic argument and not all implementation of this function
have to provide it.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{matrix}} with the
same dimensions as \code{x}.
}
\description{
Calculates the cumulative sum for each row (column) of a matrix-like object
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or
\code{\link[base]{numeric}} call \code{matrixStats::\link[matrixStats]{rowCumsums}}
/ \code{matrixStats::\link[matrixStats]{colCumsums}}.
}
\examples{
  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowCumsums(mat)
  colCumsums(mat)
  
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowCumsums}()} and
\code{matrixStats::\link[matrixStats]{colCumsums}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item \link{cumsum}
}
}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
