% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMeans.R
\name{rowMeans}
\alias{rowMeans}
\alias{rowMeans,matrix_OR_array_OR_table_OR_numeric-method}
\alias{rowMeans,data.frame-method}
\alias{rowMeans,ANY-method}
\alias{colMeans}
\alias{colMeans,matrix_OR_array_OR_table_OR_numeric-method}
\alias{colMeans,data.frame-method}
\alias{colMeans,ANY-method}
\title{Calculates the mean for each row (column) of a matrix-like object}
\usage{
rowMeans(x, na.rm = FALSE, dims = 1, ...)

\S4method{rowMeans}{matrix_OR_array_OR_table_OR_numeric}(x, na.rm = FALSE,
  dims = 1, ...)

\S4method{rowMeans}{data.frame}(x, na.rm = FALSE, dims = 1, ...)

\S4method{rowMeans}{ANY}(x, na.rm = FALSE, dims = 1, ...)

colMeans(x, na.rm = FALSE, dims = 1, ...)

\S4method{colMeans}{matrix_OR_array_OR_table_OR_numeric}(x, na.rm = FALSE,
  dims = 1, ...)

\S4method{colMeans}{data.frame}(x, na.rm = FALSE, dims = 1, ...)

\S4method{colMeans}{ANY}(x, na.rm = FALSE, dims = 1, ...)
}
\arguments{
\item{x}{An NxK matrix-like object, a numeric data frame, or an array-like
object of two or more dimensions.}

\item{na.rm}{If \code{\link[base]{TRUE}}, missing values
(\code{\link[base]{NA}} or \code{\link[base]{NaN}}) are
omitted from the calculations.}

\item{dims}{A single integer indicating which dimensions are regarded
as rows or columns to mean over. For \code{rowMeans}, the mean is over
dimensions \code{dims+1, ...}; for \code{colMeans} it is over
dimensions \code{1:dims}.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link{numeric}} \code{\link{vector}} of length N (K).
}
\description{
Calculates the mean for each row (column) of a matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link{matrix}},
\code{\link{array}}, \code{\link{table}}, \code{\link{numeric}}, or
\code{\link{data.frame}} call \code{base::\link[base]{rowMeans}} /
\code{base::\link[base]{colMeans}}.
}
\examples{
mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
mat[2, 1] <- NA
mat[3, 3] <- Inf
mat[4, 1] <- 0

print(mat)

rowMeans(mat)
colMeans(mat)
}
\seealso{
\itemize{
\item \code{base::\link[base]{colMeans}} for the default
\code{rowMeans} and \code{colMeans} methods.
\item \code{Matrix::\link[Matrix]{colMeans}} in the \pkg{Matrix} package
for \code{rowMeans} and \code{colMeans} methods defined for
CsparseMatrix derivatives (e.g. dgCMatrix objects).
}
}
\keyword{algebra}
\keyword{arith}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
