% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMeans2.R
\docType{methods}
\name{rowMeans2}
\alias{rowMeans2}
\alias{rowMeans2,matrix-method}
\alias{rowMeans2,numeric-method}
\alias{colMeans2}
\alias{colMeans2,matrix-method}
\alias{colMeans2,numeric-method}
\title{Calculates the mean for each row (column) of a matrix-like object}
\usage{
rowMeans2(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{rowMeans2}{matrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, dim. = dim(x))

\S4method{rowMeans2}{numeric}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, dim. = dim(x))

colMeans2(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{colMeans2}{matrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, dim. = dim(x))

\S4method{colMeans2}{numeric}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, dim. = dim(x))
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows, cols}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, essential when x is a
\code{\link[base]{numeric}} vector.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the mean for each row (column) of a matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or
\code{\link[base]{numeric}} call \code{matrixStats::\link[matrixStats]{rowMeans2}}
/ \code{matrixStats::\link[matrixStats]{colMeans2}}.
}
\examples{
  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowMeans2(mat)
  colMeans2(mat)
  
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowMeans2}()} and
  \code{matrixStats::\link[matrixStats]{colMeans2}()} which are used when
  the input is a \code{matrix} or \code{numeric} vector.
\item See also \code{\link[base:colSums]{rowMeans}()} for the
  corresponding function in base R.
\item For variance estimates, see \code{\link{rowVars}()}.
}
}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
