% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMedians.R
\docType{data}
\name{rowMedians}
\alias{rowMedians}
\alias{colMedians}
\title{Calculates the median for each row (column) of a matrix-like object}
\format{An object of class \code{NULL} of length 0.}
\usage{
rowMedians(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x), ...)

colMedians(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x), ...)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows, cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, also when not a
\code{\link[base]{matrix}}.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the median for each row (column) of a matrix-like object.
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowMedians}()} and
  \code{matrixStats::\link[matrixStats]{colMedians}()} for the default
  methods.
\item See \code{\link{rowWeightedMedians}()} and \code{colWeightedMedians()}
  for weighted medians.
}
}
\keyword{array}
\keyword{datasets}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
