% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowProds.R
\name{rowProds}
\alias{rowProds}
\alias{rowProds,matrix-method}
\alias{rowProds,numeric-method}
\alias{rowProds,array-method}
\alias{colProds}
\alias{colProds,matrix-method}
\alias{colProds,numeric-method}
\alias{colProds,array-method}
\title{Calculates the product for each row (column) of a matrix-like object}
\usage{
rowProds(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{rowProds}{matrix}(x, rows = NULL, cols = NULL, na.rm = FALSE,
  method = c("direct", "expSumLog"))

\S4method{rowProds}{numeric}(x, rows = NULL, cols = NULL, na.rm = FALSE,
  method = c("direct", "expSumLog"))

\S4method{rowProds}{array}(x, rows = NULL, cols = NULL, na.rm = FALSE,
  method = c("direct", "expSumLog"))

colProds(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{colProds}{matrix}(x, rows = NULL, cols = NULL, na.rm = FALSE,
  method = c("direct", "expSumLog"))

\S4method{colProds}{numeric}(x, rows = NULL, cols = NULL, na.rm = FALSE,
  method = c("direct", "expSumLog"))

\S4method{colProds}{array}(x, rows = NULL, cols = NULL, na.rm = FALSE,
  method = c("direct", "expSumLog"))
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows, cols}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}

\item{method}{A character vector of length one that specifies the
how the product is calculated. Note, that this is not a generic
argument and not all implementation have to provide it.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the product for each row (column) of a matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}},
\code{\link[base]{array}}, or \code{\link[base]{numeric}} call
\code{matrixStats::\link[matrixStats]{rowProds}}
/ \code{matrixStats::\link[matrixStats]{colProds}}.
}
\examples{
  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowProds(mat)
  colProds(mat)
  
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowProds}()} and
\code{matrixStats::\link[matrixStats]{colProds}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item For sums across rows (columns), see
\code{\link{rowSums2}()} (\code{\link[=colSums2]{colSums2()}})
\item \code{base::\link[base]{prod}()}.
}
}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
