% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowSds.R
\name{rowSds}
\alias{rowSds}
\alias{rowSds,matrix-method}
\alias{rowSds,numeric-method}
\alias{rowSds,array-method}
\alias{colSds}
\alias{colSds,matrix-method}
\alias{colSds,numeric-method}
\alias{colSds,array-method}
\title{Calculates the standard deviation for each row (column) of a matrix-like object}
\usage{
rowSds(x, rows = NULL, cols = NULL, na.rm = FALSE, center = NULL, ...)

\S4method{rowSds}{matrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  dim. = dim(x)
)

\S4method{rowSds}{numeric}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  dim. = dim(x)
)

\S4method{rowSds}{array}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  dim. = dim(x)
)

colSds(x, rows = NULL, cols = NULL, na.rm = FALSE, center = NULL, ...)

\S4method{colSds}{matrix}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  dim. = dim(x)
)

\S4method{colSds}{numeric}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  dim. = dim(x)
)

\S4method{colSds}{array}(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  center = NULL,
  dim. = dim(x)
)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows, cols}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{center}{(optional) the center, defaults to the row means}

\item{...}{Additional arguments passed to specific methods.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, essential when x is a
\code{\link[base]{numeric}} vector.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the standard deviation for each row (column) of a matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or
\code{\link[base]{numeric}} call \code{matrixStats::\link[matrixStats]{rowSds}}
/ \code{matrixStats::\link[matrixStats]{colSds}}.
}
\examples{
  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowSds(mat)
  colSds(mat)
  
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowSds}()} and
\code{matrixStats::\link[matrixStats]{colSds}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item For mean estimates, see \code{\link{rowMeans2}()} and
\code{\link[base:colSums]{rowMeans}()}.
\item For variance estimates, see \code{\link{rowVars}()}.
}
}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
