% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowVarDiffs.R
\name{rowVarDiffs}
\alias{rowVarDiffs}
\alias{rowVarDiffs,matrix-method}
\alias{rowVarDiffs,numeric-method}
\alias{rowVarDiffs,array-method}
\alias{colVarDiffs}
\alias{colVarDiffs,matrix-method}
\alias{colVarDiffs,numeric-method}
\alias{colVarDiffs,array-method}
\title{Calculates the variance of the difference between each element of a row
(column) of a matrix-like object}
\usage{
rowVarDiffs(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  diff = 1L,
  trim = 0,
  ...
)

\S4method{rowVarDiffs}{matrix}(x, rows = NULL, cols = NULL, na.rm = FALSE, diff = 1L, trim = 0)

\S4method{rowVarDiffs}{numeric}(x, rows = NULL, cols = NULL, na.rm = FALSE, diff = 1L, trim = 0)

\S4method{rowVarDiffs}{array}(x, rows = NULL, cols = NULL, na.rm = FALSE, diff = 1L, trim = 0)

colVarDiffs(
  x,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  diff = 1L,
  trim = 0,
  ...
)

\S4method{colVarDiffs}{matrix}(x, rows = NULL, cols = NULL, na.rm = FALSE, diff = 1L, trim = 0)

\S4method{colVarDiffs}{numeric}(x, rows = NULL, cols = NULL, na.rm = FALSE, diff = 1L, trim = 0)

\S4method{colVarDiffs}{array}(x, rows = NULL, cols = NULL, na.rm = FALSE, diff = 1L, trim = 0)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows, cols}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{diff}{An integer specifying the order of difference.}

\item{trim}{A double in [0,1/2] specifying the fraction of observations to
be trimmed from each end of (sorted) x before estimation.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the variance of the difference between each element of a row
(column) of a matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}},
\code{\link[base]{array}}, or \code{\link[base]{numeric}} call
\code{matrixStats::\link[matrixStats]{rowVarDiffs}}
/ \code{matrixStats::\link[matrixStats]{colVarDiffs}}.
}
\examples{
  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowVarDiffs(mat)
  colVarDiffs(mat)
  
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowVarDiffs}()} and
\code{matrixStats::\link[matrixStats]{colVarDiffs}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item for the direct variance see \code{\link[=rowVars]{rowVars()}}.
}
}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
