% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowWeightedMads.R
\name{rowWeightedMads}
\alias{rowWeightedMads}
\alias{rowWeightedMads,matrix-method}
\alias{rowWeightedMads,numeric-method}
\alias{rowWeightedMads,array-method}
\alias{colWeightedMads}
\alias{colWeightedMads,matrix-method}
\alias{colWeightedMads,numeric-method}
\alias{colWeightedMads,array-method}
\title{Calculates the weighted median absolute deviation for each row (column) of a matrix-like object}
\usage{
rowWeightedMads(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  constant = 1.4826,
  center = NULL,
  ...
)

\S4method{rowWeightedMads}{matrix}(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  constant = 1.4826,
  center = NULL
)

\S4method{rowWeightedMads}{numeric}(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  constant = 1.4826,
  center = NULL
)

\S4method{rowWeightedMads}{array}(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  constant = 1.4826,
  center = NULL
)

colWeightedMads(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  constant = 1.4826,
  center = NULL,
  ...
)

\S4method{colWeightedMads}{matrix}(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  constant = 1.4826,
  center = NULL
)

\S4method{colWeightedMads}{numeric}(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  constant = 1.4826,
  center = NULL
)

\S4method{colWeightedMads}{array}(
  x,
  w = NULL,
  rows = NULL,
  cols = NULL,
  na.rm = FALSE,
  constant = 1.4826,
  center = NULL
)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{w}{A \code{\link{numeric}} vector of length K (N) that specifies by
how much each element is weighted.}

\item{rows, cols}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{constant}{A scale factor. See \link[base:mad]{base::mad} for details.}

\item{center}{(optional) the center, defaults to the row means}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the weighted  median absolute deviation for each row (column) of a matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or
\code{\link[base]{numeric}} call \code{matrixStats::\link[matrixStats]{rowWeightedMads}}
/ \code{matrixStats::\link[matrixStats]{colWeightedMads}}.
}
\examples{
  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  w <- rnorm(n = 5, mean = 3)
  rowWeightedMads(mat, w = w[1:3])
  colWeightedMads(mat, w = w)
  
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowWeightedMads}()} and
\code{matrixStats::\link[matrixStats]{colWeightedMads}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item See also \link{rowMads} for the corresponding unweighted function.
}
}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
