% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowWeightedMedians.R
\docType{data}
\name{rowWeightedMedians}
\alias{rowWeightedMedians}
\alias{colWeightedMedians}
\title{Calculates the weighted medians for each row (column) in a matrix-like object}
\format{An object of class \code{NULL} of length 0.}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{w}{A \code{\link[base]{numeric}} \code{\link[base]{vector}} of length
K (N).}

\item{rows, cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, missing values are
excluded from the calculation, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the weighted medians for each row (column) in a matrix-like object.
}
\seealso{
\itemize{
\item See \code{\link[matrixStats]{rowWeightedMedians}()} and
  \code{\link[matrixStats]{colWeightedMedians}()} for the default methods.
\item See \code{\link{rowMedians}()} and \code{colMedians()} for
  non-weighted medians.
}
}
\keyword{array}
\keyword{datasets}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
