% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowAnyNAs.R
\name{rowAnyNAs}
\alias{rowAnyNAs}
\alias{rowAnyNAs,matrix-method}
\alias{rowAnyNAs,numeric-method}
\alias{rowAnyNAs,array-method}
\alias{colAnyNAs}
\alias{colAnyNAs,matrix-method}
\alias{colAnyNAs,numeric-method}
\alias{colAnyNAs,array-method}
\title{Check if any elements in a row (column) of a matrix-like object is missing}
\usage{
rowAnyNAs(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{rowAnyNAs}{matrix}(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x))

\S4method{rowAnyNAs}{numeric}(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x))

\S4method{rowAnyNAs}{array}(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x))

colAnyNAs(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{colAnyNAs}{matrix}(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x))

\S4method{colAnyNAs}{numeric}(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x))

\S4method{colAnyNAs}{array}(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x))
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows, cols}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, essential when x is a
\code{\link[base]{numeric}} vector.}
}
\value{
Returns a \code{\link[base]{logical}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Check if any elements in a row (column) of a matrix-like object is missing
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or
\code{\link[base]{numeric}} call \code{matrixStats::\link[matrixStats]{rowAnyNAs}}
/ \code{matrixStats::\link[matrixStats]{colAnyNAs}}.
}
\examples{
  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowAnyNAs(mat)
  colAnyNAs(mat)
  
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowAnyNAs}()} and
  \code{matrixStats::\link[matrixStats]{colAnyNAs}()} which are used when
  the input is a \code{matrix} or \code{numeric} vector.
\item For checks if any element is equal to a value, see \code{\link{rowAnyNAs}()}.
}
}
