% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowDiffs.R
\name{rowDiffs}
\alias{rowDiffs}
\alias{rowDiffs,matrix-method}
\alias{rowDiffs,numeric-method}
\alias{rowDiffs,array-method}
\alias{colDiffs}
\alias{colDiffs,matrix-method}
\alias{colDiffs,numeric-method}
\alias{colDiffs,array-method}
\title{Calculates the difference between each element of a row (column) of a matrix-like object}
\usage{
rowDiffs(x, rows = NULL, cols = NULL, lag = 1L, differences = 1L, ...)

\S4method{rowDiffs}{matrix}(
  x,
  rows = NULL,
  cols = NULL,
  lag = 1L,
  differences = 1L,
  dim. = dim(x)
)

\S4method{rowDiffs}{numeric}(
  x,
  rows = NULL,
  cols = NULL,
  lag = 1L,
  differences = 1L,
  dim. = dim(x)
)

\S4method{rowDiffs}{array}(
  x,
  rows = NULL,
  cols = NULL,
  lag = 1L,
  differences = 1L,
  dim. = dim(x)
)

colDiffs(x, rows = NULL, cols = NULL, lag = 1L, differences = 1L, ...)

\S4method{colDiffs}{matrix}(
  x,
  rows = NULL,
  cols = NULL,
  lag = 1L,
  differences = 1L,
  dim. = dim(x)
)

\S4method{colDiffs}{numeric}(
  x,
  rows = NULL,
  cols = NULL,
  lag = 1L,
  differences = 1L,
  dim. = dim(x)
)

\S4method{colDiffs}{array}(
  x,
  rows = NULL,
  cols = NULL,
  lag = 1L,
  differences = 1L,
  dim. = dim(x)
)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows, cols}{A \code{\link[base]{vector}} indicating the subset (and/or
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{lag}{An integer specifying the lag.}

\item{differences}{An integer specifying the order of difference.}

\item{...}{Additional arguments passed to specific methods.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, essential when x is a
\code{\link[base]{numeric}} vector.
Note, that this is not a generic argument and not all implementation of this function
have to provide it.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{matrix}} with one
column (row) less than x: \eqn{Nx(K-1)} or \eqn{(N-1)xK}.
}
\description{
Calculates the difference between each element of a row (column) of a matrix-like object
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or
\code{\link[base]{numeric}} call \code{matrixStats::\link[matrixStats]{rowDiffs}}
/ \code{matrixStats::\link[matrixStats]{colDiffs}}.
}
\examples{
  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowDiffs(mat)
  colDiffs(mat)
  
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowDiffs}()} and
\code{matrixStats::\link[matrixStats]{colDiffs}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item \link{diff}
}
}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
