% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMedians.R
\docType{methods}
\name{rowMedians}
\alias{rowMedians}
\alias{rowMedians,matrix-method}
\alias{rowMedians,numeric-method}
\alias{colMedians}
\alias{colMedians,matrix-method}
\alias{colMedians,numeric-method}
\title{Calculates the median for each row (column) of a matrix-like object}
\usage{
rowMedians(x, na.rm = FALSE, ...)

\S4method{rowMedians}{matrix}(x, na.rm = FALSE, ...)

\S4method{rowMedians}{numeric}(x, na.rm = FALSE, dim., ...)

colMedians(x, na.rm = FALSE, ...)

\S4method{colMedians}{matrix}(x, na.rm = FALSE, ...)

\S4method{colMedians}{numeric}(x, na.rm = FALSE, dim., ...)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, important when x is a
\code{\link[base]{numeric}} vector.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the median for each row (column) of a matrix-like object.
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowMedians}()} and
  \code{matrixStats::\link[matrixStats]{colMedians}()} which are used for
  \code{matrix} input.
\item See \code{\link{rowWeightedMedians}()} and \code{colWeightedMedians()}
  for weighted medians.
}
}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
