% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMedians.R
\docType{methods}
\name{rowMedians}
\alias{rowMedians}
\alias{rowMedians,matrix-method}
\alias{rowMedians,numeric-method}
\alias{colMedians}
\alias{colMedians,matrix-method}
\alias{colMedians,numeric-method}
\title{Calculates the median for each row (column) of a matrix-like object}
\usage{
rowMedians(x, na.rm = FALSE, ...)

\S4method{rowMedians}{matrix}(x, na.rm = FALSE, rows = NULL,
  cols = NULL, dim. = dim(x))

\S4method{rowMedians}{numeric}(x, na.rm = FALSE, rows = NULL,
  cols = NULL, dim. = dim(x))

colMedians(x, na.rm = FALSE, ...)

\S4method{colMedians}{matrix}(x, na.rm = FALSE, rows = NULL,
  cols = NULL, dim. = dim(x))

\S4method{colMedians}{numeric}(x, na.rm = FALSE, rows = NULL,
  cols = NULL, dim. = dim(x))
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}

\item{rows, cols}{A \code{\link[base]{vector}} indicating the subset (and/or 
columns) to operate over. If \code{\link[base]{NULL}}, no subsetting is done.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, essential when x is a
\code{\link[base]{numeric}} vector.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the median for each row (column) of a matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}} or 
\code{\link[base]{numeric}} are dispatched to \code{matrixStats::\link[matrixStats]{rowMedians}}

Methods that provide an implemention for the S4 for generic are encouraged
to handle the \code{rows}/code{cols} subsetting.
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowMedians}()} and
  \code{matrixStats::\link[matrixStats]{colMedians}()} which are used for
  \code{matrix} input.
\item See \code{\link{rowWeightedMedians}()} and \code{colWeightedMedians()}
  for weighted medians.
}
}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
