% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMedians.R
\name{rowMedians}
\alias{rowMedians}
\alias{colMedians}
\title{Calculates the median for each row (column) of a matrix-like object}
\usage{
rowMedians(x, na.rm = FALSE, ...)

colMedians(x, na.rm = FALSE, ...)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the median for each row (column) of a matrix-like object.
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowMedians}()} and
  \code{matrixStats::\link[matrixStats]{colMedians}()} which are used for
  \code{matrix} input.
\item See \code{\link{rowWeightedMedians}()} and \code{colWeightedMedians()}
  for weighted medians.
}
}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
