% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowMins.R
\name{rowMins}
\alias{rowMins}
\alias{rowMins,matrix-method}
\alias{rowMins,numeric-method}
\alias{rowMins,array-method}
\alias{colMins}
\alias{colMins,matrix-method}
\alias{colMins,numeric-method}
\alias{colMins,array-method}
\title{Calculates the minimum for each row (column) of a matrix-like object}
\usage{
rowMins(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{rowMins}{matrix}(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x))

\S4method{rowMins}{numeric}(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x))

\S4method{rowMins}{array}(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x))

colMins(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{colMins}{matrix}(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x))

\S4method{colMins}{numeric}(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x))

\S4method{colMins}{array}(x, rows = NULL, cols = NULL, na.rm = FALSE, dim. = dim(x))
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{rows, cols}{A \code{\link[base]{vector}} indicating the subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no
subsetting is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are excluded first, otherwise not.}

\item{...}{Additional arguments passed to specific methods.}

\item{dim.}{An \code{\link[base]{integer}} \code{\link[base]{vector}} of
length two specifying the dimension of \code{x}, essential when x is a
\code{\link[base]{numeric}} vector. Note, that this is not a generic
argument and not all methods need provide it.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the minimum for each row (column) of a matrix-like object.
}
\details{
The S4 methods for \code{x} of type \code{\link[base]{matrix}},
\code{\link[base]{array}}, or \code{\link[base]{numeric}} call
\code{matrixStats::\link[matrixStats]{rowMins}}
/ \code{matrixStats::\link[matrixStats]{colMins}}.
}
\examples{
  mat <- matrix(rnorm(15), nrow = 5, ncol = 3)
  mat[2, 1] <- NA
  mat[3, 3] <- Inf
  mat[4, 1] <- 0
  
  print(mat)
  
  rowMins(mat)
  colMins(mat)
  
}
\seealso{
\itemize{
\item \code{matrixStats::\link[matrixStats]{rowMins}()} and
\code{matrixStats::\link[matrixStats]{colMins}()} which are used when
the input is a \code{matrix} or \code{numeric} vector.
\item For max estimates, see \code{\link{rowMaxs}()}.
}
}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
