% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowWeightedMedians.R
\docType{methods}
\name{rowWeightedMedians}
\alias{rowWeightedMedians}
\alias{rowWeightedMedians,matrix-method}
\alias{colWeightedMedians}
\alias{colWeightedMedians,matrix-method}
\title{Calculates the weighted median for each row (column) of a matrix-like object}
\usage{
rowWeightedMedians(x, w = NULL, rows = NULL, cols = NULL,
  na.rm = FALSE, ...)

\S4method{rowWeightedMedians}{matrix}(x, w = NULL, rows = NULL,
  cols = NULL, na.rm = FALSE, ...)

colWeightedMedians(x, rows = NULL, cols = NULL, na.rm = FALSE, ...)

\S4method{colWeightedMedians}{matrix}(x, rows = NULL, cols = NULL,
  na.rm = FALSE, ...)
}
\arguments{
\item{x}{An NxK matrix-like object.}

\item{w}{A \code{\link[base]{numeric}} \code{\link[base]{vector}} of length
K (N).}

\item{rows, cols}{A \code{\link[base]{vector}} indicating subset of rows
(and/or columns) to operate over. If \code{\link[base]{NULL}}, no subsetting
is done.}

\item{na.rm}{If \code{\link[base:logical]{TRUE}}, \code{\link[base]{NA}}s
are stripped from x or not.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Returns a \code{\link[base]{numeric}} \code{\link[base]{vector}} of
length N (K).
}
\description{
Calculates the weighted median for each row (column) of a matrix-like object.
}
\seealso{
\itemize{
\item See \code{matrixStats::\link[matrixStats]{rowWeightedMedians}()} and
  \code{matrixStats::\link[matrixStats]{colWeightedMedians}()} which are used for
  \code{matrix} input.
\item See \code{\link{rowMedians}()} and \code{colMedians()}
  for classical medians.
}
}
\keyword{array}
\keyword{iteration}
\keyword{robust}
\keyword{univar}
